Address PR feedback for #1143
