+++ frontend/src/entrypoints/proposals.tsx
@@ -4,9 +4,7 @@ import { z } from 'zod';
 import { mountPage } from '../boot/mountPage';
 import { BootPayload } from '../boot/parseBootPayload';
 import { executionStatusPillClasses } from '../utils/executionStatusPillClasses';
-
-const PAGE_SIZE_OPTIONS = [20, 25, 50, 100] as const;
-const DEFAULT_PAGE_SIZE = 50;
+import { PageSizeSelector, parsePageSize } from '../components/PageSizeSelector';
 
 const PROPOSAL_STATUSES = ['open', 'promoted', 'dismissed'] as const;
 
@@ -36,15 +34,6 @@ const ProposalsResponseSchema = z.object({
   items: z.array(ProposalSchema),
   nextCursor: z.string().nullable().optional(),
 });
-
-
-
-function parsePageSize(raw: string | null): number {
-  const parsed = Number(raw || DEFAULT_PAGE_SIZE);
-  return PAGE_SIZE_OPTIONS.includes(parsed as (typeof PAGE_SIZE_OPTIONS)[number])
-    ? parsed
-    : DEFAULT_PAGE_SIZE;
-}
 
 function formatWhen(iso: string | null | undefined): string {
   if (!iso) return '—';
@@ -244,43 +233,35 @@ function ProposalsPage({ payload }: { payload: BootPayload }) {
         <div className="queue-layouts">
           <div className="queue-results-toolbar">
             <span className="small">{pageSummary}</span>
-            <nav className="queue-pagination" aria-label="Pagination">
-              <label className="queue-inline-filter">
-                Show
-                <select
-                  value={pageSize}
-                  onChange={(event) => {
-                    setPageSize(parsePageSize(event.target.value));
-                    resetToFirstPage();
-                  }}
-                  aria-label="Rows per page"
+            <div className="queue-pagination">
+              <PageSizeSelector
+                pageSize={pageSize}
+                onPageSizeChange={(size) => {
+                  setPageSize(size);
+                  resetToFirstPage();
+                }}
+              />
+              <nav aria-label="Pagination" style={{ display: 'inline-flex', gap: '0.45rem' }}>
+                <button
+                  type="button"
+                  className="secondary queue-pagination-button"
+                  disabled={cursorStack.length === 0}
+                  onClick={goPrev}
+                  aria-label="Previous page"
                 >
-                  {PAGE_SIZE_OPTIONS.map((size) => (
-                    <option key={size} value={size}>
-                      {size}
-                    </option>
-                  ))}
-                </select>
-              </label>
-              <button
-                type="button"
-                className="secondary queue-pagination-button"
-                disabled={cursorStack.length === 0}
-                onClick={goPrev}
-                aria-label="Previous page"
-              >
-                <span aria-hidden="true">&larr;</span>
-              </button>
-              <button
-                type="button"
-                className="secondary queue-pagination-button"
-                disabled={!data?.nextCursor}
-                onClick={goNext}
-                aria-label="Next page"
-              >
-                <span aria-hidden="true">&rarr;</span>
-              </button>
-            </nav>
+                  <span aria-hidden="true">&larr;</span>
+                </button>
+                <button
+                  type="button"
+                  className="secondary queue-pagination-button"
+                  disabled={!data?.nextCursor}
+                  onClick={goNext}
+                  aria-label="Next page"
+                >
+                  <span aria-hidden="true">&rarr;</span>
+                </button>
+              </nav>
+            </div>
           </div>
           {sortedItems.length === 0 ? (
             <div className="card small">No proposals found for the current filters.</div>
