+++ frontend/src/entrypoints/tasks-list.tsx
@@ -5,9 +5,8 @@ import { z } from 'zod';
 import { mountPage } from '../boot/mountPage';
 import { BootPayload } from '../boot/parseBootPayload';
 import { executionStatusPillClasses } from '../utils/executionStatusPillClasses';
+import { PageSizeSelector, parsePageSize } from '../components/PageSizeSelector';
 
-const PAGE_SIZE_OPTIONS = [20, 25, 50, 100] as const;
-const DEFAULT_PAGE_SIZE = 50;
 const POLL_MS_DEFAULT = 5000;
 
 type ListDashboardConfig = {
@@ -86,13 +85,6 @@ function readListDashboardConfig(payload: BootPayload): ListDashboardConfig | un
 function normalizeTableSortField(raw: string | null): string {
   const candidate = (raw || '').trim();
   return VALID_TABLE_SORT_FIELDS.has(candidate) ? candidate : 'createdAt';
-}
-
-function parsePageSize(raw: string | null): number {
-  const parsed = Number(raw || DEFAULT_PAGE_SIZE);
-  return PAGE_SIZE_OPTIONS.includes(parsed as (typeof PAGE_SIZE_OPTIONS)[number])
-    ? parsed
-    : DEFAULT_PAGE_SIZE;
 }
 
 function formatWhen(iso: string | null | undefined): string {
@@ -414,44 +406,36 @@ export function TasksListPage({ payload }: { payload: BootPayload }) {
         <div className="queue-layouts">
           <div className="queue-results-toolbar">
             <span className="small">{pageSummary}</span>
-            <nav className="queue-pagination" aria-label="Pagination">
-              <label className="queue-inline-filter">
-                Show
-                <select
-                  value={pageSize}
-                  disabled={!listEnabled}
-                  onChange={(event) => {
-                    setPageSize(parsePageSize(event.target.value));
-                    resetToFirstPage();
-                  }}
-                  aria-label="Rows per page"
+            <div className="queue-pagination">
+              <PageSizeSelector
+                pageSize={pageSize}
+                disabled={!listEnabled}
+                onPageSizeChange={(size) => {
+                  setPageSize(size);
+                  resetToFirstPage();
+                }}
+              />
+              <nav aria-label="Pagination" style={{ display: 'inline-flex', gap: '0.45rem' }}>
+                <button
+                  type="button"
+                  className="secondary queue-pagination-button"
+                  disabled={!listEnabled || cursorStack.length === 0}
+                  onClick={goPrev}
+                  aria-label="Previous page"
                 >
-                  {PAGE_SIZE_OPTIONS.map((size) => (
-                    <option key={size} value={size}>
-                      {size}
-                    </option>
-                  ))}
-                </select>
-              </label>
-              <button
-                type="button"
-                className="secondary queue-pagination-button"
-                disabled={!listEnabled || cursorStack.length === 0}
-                onClick={goPrev}
-                aria-label="Previous page"
-              >
-                <span aria-hidden="true">&larr;</span>
-              </button>
-              <button
-                type="button"
-                className="secondary queue-pagination-button"
-                disabled={!listEnabled || !data?.nextPageToken}
-                onClick={goNext}
-                aria-label="Next page"
-              >
-                <span aria-hidden="true">&rarr;</span>
-              </button>
-            </nav>
+                  <span aria-hidden="true">&larr;</span>
+                </button>
+                <button
+                  type="button"
+                  className="secondary queue-pagination-button"
+                  disabled={!listEnabled || !data?.nextPageToken}
+                  onClick={goNext}
+                  aria-label="Next page"
+                >
+                  <span aria-hidden="true">&rarr;</span>
+                </button>
+              </nav>
+            </div>
           </div>
           {sortedItems.length === 0 ? (
             <div className="card small">No tasks found for the current filters.</div>
